+++ src/gpu.jl
@@ -1,15 +1,19 @@
-macro cuda1(lib,fun,x...)       # quietly do nothing if library missing
+macro gpu(_ex); if gpu()>=0; esc(_ex); end; end
+
+macro cuda1(lib,fun,x...)       # return -1 if library missing, error code if run
+    if Libdl.find_library(["lib$lib"], []) != ""
+        Expr(:ccall, ("$fun","lib$lib"), :UInt32, x...)
+    else
+        -1
+    end
+end
+
+macro cuda(lib,fun,x...)        # give an error if library missing, warning if return!=0
     if Libdl.find_library(["lib$lib"], []) != ""
         fx = Expr(:ccall, ("$fun","lib$lib"), :UInt32, x...)
         msg = "$lib.$fun error "
         err = gensym()
         esc(:(if ($err=$fx) != 0; warn($msg, $err); Base.show_backtrace(STDOUT, backtrace()); end))
-    end
-end
-
-macro cuda(lib,fun,x...)        # give an error if library missing
-    if Libdl.find_library(["lib$lib"], []) != ""
-        Expr(:macrocall,Symbol("@cuda1"),lib,fun,x...)
     else
         Expr(:call,:error,"Cannot find lib$lib, please install it and rerun Pkg.build(\"Knet\").")
     end