+++ test/update.jl
@@ -1,80 +1,82 @@
 using Knet, Base.Test
+srand(123456789)
 
 # x* = f(1, · · · , 1)
 # f(x*) = 0
 
 rosenbrock(x) = sum((1-x[1:end-1]).^2 + 100*(x[2:end]-x[1:end-1].^2).^2)
 
-grads = grad(rosenbrock)
-srand(123456789)
+grads = gradloss(rosenbrock)
+dims = 10
+ftol = 1e-3
+xtol = 1e-10
+maxiter = 25000
 
 function test_base(w, params)
-	prev = 0
-	i = 1
-
-	current = rosenbrock(w)
-	while i <= 1000000 && abs(current - prev) > 1e-10 && current > 1e-3
-		g = grads(w)
-		w, params = update!(w, g, params)
-		prev = current
-		current = rosenbrock(w)
-
-		i += 1
-	end
-
-	@test current <= 1e-3
-	info("$(typeof(params)) Passed!\nConverged to $(current) at epoch $(i-1)")
+    i = 1
+    prev = 0
+    current = 1
+    while i <= maxiter && abs(current - prev) > xtol && current > ftol
+	prev = current
+	g, current = grads(w)
+	w, params  = update!(w, g, params)
+	i += 1
+    end
+    @test current <= ftol
+    info("$(typeof(params)) for $(typeof(w)) $(current <= ftol ? "Passed" : "Failed")!\nConverged to $(current) at epoch $(i-1)")
 end
 
 function test_sgd(w)
-	params = Sgd(;lr=0.0005)
-	test_base(w, params)
+    params = Sgd(;lr=0.0005)
+    test_base(w, params)
 end
 
 function test_momentum(w)
-	params = Momentum(w;lr=0.00005, gamma=0.95, velocity=zeros(w))
-	test_base(w, params)
+    params = Momentum(w;lr=0.00025, gamma=0.95)
+    test_base(w, params)
 end
 
 function test_adam(w)
-	params = Adam(w; lr=0.005, beta1=0.9, beta2=0.95, t=1, eps=1e-8, fstm=zeros(w), scndm=zeros(w))
-	test_base(w, params)
+    params = Adam(w; lr=0.005, beta1=0.9, beta2=0.95, t=1, eps=1e-8)
+    test_base(w, params)
 end
 
 function test_adagrad(w)
-	params = Adagrad(w; lr=0.35, eps=1e-6, G=zeros(w))
-	test_base(w, params)
+    params = Adagrad(w; lr=0.35, eps=1e-6)
+    test_base(w, params)
 end
 
 function test_adadelta(w)
-	params = Adadelta(w; lr=0.001, rho=0.9, eps=1e-6, G=zeros(w), delta=zeros(w))
-	test_base(w, params)
+    params = Adadelta(w; lr=0.01, rho=0.5, eps=1e-6)
+    test_base(w, params)
 end
 
 function test_rmsprop(w)
-	params = Rmsprop(w; lr=0.0002, rho=0.9, eps=1e-6, G=zeros(w))
-	test_base(w, params)
+    params = Rmsprop(w; lr=0.0005, rho=0.9, eps=1e-6)
+    test_base(w, params)
 end
 
-#GPU Tests
-info("GPU Tests")
-raw = randn(100, 1)
-w = convert(KnetArray, raw)
-@time test_sgd(copy(w))
-@time test_momentum(copy(w))
-@time test_adam(copy(w))
-@time test_adagrad(copy(w))
-@time test_adadelta(copy(w))
-@time test_rmsprop(copy(w))
+@testset "update!" begin
 
 #CPU Tests
-gpu(-1)
-info("CPU Tests")
-w = randn(100, 1)
-@time test_sgd(copy(w))
-@time test_momentum(copy(raw))
-@time test_adam(copy(raw))
-@time test_adagrad(copy(raw))
-@time test_adadelta(copy(raw))
-@time test_rmsprop(copy(raw))
+    w = randn(dims)
+    @time test_sgd(copy(w))
+    @time test_momentum(copy(w))
+    @time test_adam(copy(w))
+    @time test_adagrad(copy(w))
+    @time test_adadelta(copy(w))
+    @time test_rmsprop(copy(w))
 
+#GPU Tests
+    if gpu() >= 0
+        w = KnetArray(w)
+        @time test_sgd(copy(w))
+        @time test_momentum(copy(w))
+        @time test_adam(copy(w))
+        @time test_adagrad(copy(w))
+        @time test_adadelta(copy(w))
+        @time test_rmsprop(copy(w))
+    end
+end
+
+nothing